refactor(onboard): extract host-qr/wechat/channel-setup to onboard/ helpers
Brings src/lib/onboard.ts under the per-file growth budget enforced by
the source-shape check (net -93 lines vs origin/main on this PR; was
+129 before). New logic moved into focused modules under src/lib/onboard/
instead of the top-level entrypoint:

- onboard/host-qr-dispatch.ts: dispatchHostQrLogin() runs a channel's
  host-side QR handler and applies token + non-secret metadata side
  effects. Wraps the await in a real try/catch so a non-async handler
  that throws during setup still becomes a structured "error" outcome.
- onboard/wechat-config.ts: gatherWechatConfig() reads
  WECHAT_ACCOUNT_ID/_BASE_URL/_USER_ID from env with a session fallback;
  hasWechatConfigDrift() compares fresh env vs recorded session for
  resume-time drift detection; toSessionWechatConfig() builds the
  Session.wechatConfig payload (null-on-empty for parseWechatConfig
  compatibility).
- onboard/messaging-channel-setup.ts: setupSelectedMessagingChannels()
  drives the per-channel interactive loop (token / app token / server
  ID / mention mode / allowlist IDs). The host-qr branch routes through
  the dispatcher above.

src/lib/onboard.ts now references these via single-line require()s with
typeof import() types, matching the existing helper-import pattern.

No behavior change. 223/223 onboard + dockerfile-patch tests pass.
